requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
cssselect==1.2.0
Deprecated==1.2.14
diskcache==5.6.3
einops==0.7.0
emoji==2.10.1
fake-useragent==1.5.1
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==24.3.7
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
googlesearch-python==1.2.3
grpcio==1.62.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.3.2
itsdangerous==2.1.2
Jinja2==3.1.3
kubernetes==29.0.0
llama_cpp_python==0.2.57
lxml==5.1.0
MarkupSafe==2.1.5
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.1
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==24.0
pandas==2.2.1
parse==1.20.1
pillow==10.2.0
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.1.0
PyPika==0.48.9
pyppeteer==2.0.0
pyproject_hooks==1.0.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.21.0
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
